requirements.txt
customtkinter
Pillow
pygame
opencv-python
pyautogui
pyperclip
SpeechRecognition
openai
pywin32
python-dotenv
